refactor(model-selector): unify LLMPopover, ModelSelector, LLMSelector into one
- New web/src/sections/model-selector/ with three focused components:
  · ModelSelector — single-model popover (value: model_configuration_id)
  · MultiModelSelector — multi-model pill selector (moved from refresh-components)
  · ModelSelectorContent — shared search + grouped list (moved from ModelListContent)
- New web/src/lib/languageModels/options.ts — canonical home for
  buildLlmOptions, groupLlmOptions, findModelConfigId, LLMOption, LLMOptionGroup
  (consolidated from llmUtils.ts, LLMPopover.tsx exports, and interfaces.ts)
- Delete LLMPopover.tsx, ModelSelector.tsx, ModelListContent.tsx, llmUtils.ts,
  interfaces.ts (refresh-components/popovers) and LLMSelector.tsx (components/llm)
- Migrate call sites:
  · MessageToolbar: LLMPopover foldable → ModelSelector with renderTrigger
  · SettingsPage: LLMPopover → ModelSelector with temperatureManager
  · AgentEditorPage: LLMSelector + getCurrentLlm/onLlmSelect → ModelSelector
  · AppPage + NRFPage: old ModelSelector → MultiModelSelector
  · IndexSettingsPage: ModelListContent → ModelSelectorContent
  · useMultiModelChat, useChatController, ChatUI: update import paths
