+++ web/src/refresh-pages/AppPage.tsx
@@ -34,7 +34,7 @@ import { FederatedConnectorDetail, UserRole, ValidSources } from "@/lib/types";
 import DocumentsSidebar from "@/sections/document-sidebar/DocumentsSidebar";
 import useChatController from "@/hooks/useChatController";
 import useMultiModelChat from "@/hooks/useMultiModelChat";
-import ModelSelector from "@/refresh-components/popovers/ModelSelector";
+import MultiModelSelector from "@/sections/model-selector/MultiModelSelector";
 import { useAgentController } from "@/lib/agents/hooks";
 import useChatSessionController from "@/hooks/useChatSessionController";
 import useDeepResearchToggle from "@/hooks/useDeepResearchToggle";
@@ -910,7 +910,7 @@ export default function AppPage({ firstMessage }: ChatPageProps) {
                         ) &&
                         liveAgent &&
                         !llmManager.isLoadingProviders && (
-                          <ModelSelector
+                          <MultiModelSelector
                             llmManager={llmManager}
                             selectedModels={multiModel.selectedModels}
                             onAdd={multiModel.addModel}
@@ -987,7 +987,7 @@ export default function AppPage({ firstMessage }: ChatPageProps) {
                         liveAgent &&
                         !llmManager.isLoadingProviders && (
                           <div className="pb-1">
-                            <ModelSelector
+                            <MultiModelSelector
                               llmManager={llmManager}
                               selectedModels={multiModel.selectedModels}
                               onAdd={multiModel.addModel}
